Extract Member class
Member data is spread over two classes: MemberSection and MemberPage.

Previously we were returning all data in MemberSection but some of
that data came via MemberPage and so MemberSection was representing
more than just a section of the MembersPage.

This commit creates a class to represent a single Member. It returns
a MemberSection and a MemberPage for that member.